+++ Assets/MasterServerToolkit/MasterServer/Scripts/Modules/Chat/ChatChannelInfo.cs
@@ -7,6 +7,6 @@ namespace MasterServerToolkit.MasterServer
     public class ChatChannelInfo
     {
         public string Name { get; set; }
-        public ushort OnlineCount { get; set; }
+        public int OnlineCount { get; set; }
     }
 }
